service/requirements.txt
# API
flask
flask_cors
flask_jwt_extended
requests
lxml_html_clean


# TEXT TO SPEECH
tts
torch
torchaudio

# LANGUAGE MODEL
ollama
newspaper3k
ddgs
# MACHINE LEARNING

# OTHER
coloredlogs
load_dotenv